Add configuration contrubutes Fixes #2
